Make D.JS Great Again!
The bugs, they are pouring in, we're going to deport all the bugs! We're going to build a wall, a great wall, to keep all of the bugs out!

I'm, like, a really smart person. I know codes... Believe me. I have the best codes, I have the best, but there is no better code than mine, Right?

All of 'em are weak, the other contributors are just weak. Some of them are fine people. But they are weak.

It's ok guys, because I have the answer right here! It's gonna be huge! It's going to be tremendous!! Like you'll never believe!