Update index.d.ts - Addition to Channel class
Send function added to channel class in typescript typings. Function seems to take a string and return a promise that should resolve to a Message. I'm presuming there are options to provide other argument types for sending files etc, but this was an issue that I encountered fairly early on in my usage of the library. Let me know if I'm missing something / your thoughts on this PR generally. 
Thanks for an awesome library.