docs(Permissions): Permissions Name Update
This update is for Discords new video call rollout on servers, changing the `STREAM` permission to `VIDEO` which allows both. This might be temporary as Discord is currently deciding whether to use 1 or both permissions. I recommend this PR hold off for a little while discord makes the necessary changes to accompany this rollout.