fix(GuildMemberRoleManager): fix remove method when passing collection/array
When passing a collection of roles or an array of RoleResolvable, it gets mapped into an array of IDs and then filtered but checking if the array contains the role and not the id. This commit checks if the array contains the role id, allowing GuildMemberRoleManager#remove to work with a collection or array.