Make options parameter to Client implementation class constructor optional, or possibly a blank object
Hi.  New user.  Could be wrong.

Most of the tutorials I find show code like `const client = new Discord.client();`.

[The actual constructor implementation](https://github.com/discordjs/discord.js/blob/main/src/client/Client.js#L38) defers to a class inherited constructor superclass, [which offers a blank object as a default value](https://github.com/discordjs/discord.js/blob/main/src/client/BaseClient.js#L13).

The subordinate class never actually refers to the constructor argument again after calling the superclass; instead it refers to the class members, which were set by the parent method.

On these grounds, I believe that you intend to allow a user to be allowed to pass 

1. Nothing
2. A blank object

Currently, when running in TypeScript, you are required to pass a valid instance of `ClientOptions`, which means that as a new user you have to figure out what the hell an `Intent` is.

I think this isn't intentional.

I have made the parameter optional, and I have allowed the substitution of a blank object for the argument.

Please consider.  TY.