This should solve the API hiccup instance, and allow the library to recover (ideally)
https://cdn.hyperz.net/main/4OB9Ap.png

It's a bad practice to not have common instances like this checked for when the API has a hiccup like it did a few moments ago and crashed bots that ran this function. This change will implement a fix for the above error. I am not the only person who got this error either.

Other: https://discord.com/channels/222078108977594368/824411059443204127/922550844195496047