chore: fix git cliff configs
Little while ago I remarked to @iCrawl in DM that there was a bug in the
git cliff configs that I had to fix for Sapphire. Namely when a commit
has the following body:
```
type: message

BREAKING CHANGE: One
BREAKING CHANGE: Two
... etc
```

These breaking changes weren't being listed in the changelog because of
how Git Cliff parses the commit body and footer.

That all said, this PR fixes that issue.

Lastly, when making the changes for Sapphire I thought it'd be fun to
add emojis to the scopes. Some other projects do this and it really
makes their changelogs look more fun. Please do let me know if for DiscordJS
you guys do not want the emojis and I'll remove them again.