refactor(rest): add web compatibility
This PR makes the following changes:

- `node:` prefixes for *value* imports have been removed in favor
of global objects.
- `ResponseLike` not has an additional property `statusText` as
`STATUS_CODES` is only in node and `fetch`'s statusText seems to provide
equivalent functionality
- `file-type` has been removed due to dependence on node and `is-apng`
is it's replacement
- Various fixes have been made to how the undici strategy is imported.
`undiciRequest` was being bundled even if `fetch` was specified as the
`makeRequest` option, despite it being a dynamic import. This PR fixes
that issue by manually specifying `undiciRequest.ts` should not be
bundled.