Only update name of GuildScheduledEvent if found in data
When e. g. the event "guildScheduledEventUserAdd" is fired, it runs the _patch function of the GuildScheduledEvent object with just the data parameters "id" and "guild_id".
Without checking if "name" is in the data object, the name is set to "undefined" when a user subscribes to the event and no longer retrievable unless one fetches the event again.